minor PMD, Checkstyle, and FindBugs fixes (#14)
Minor PMD, Checkstyle, and FindBugs fixes

* Removed unnecessary warning suppression

* Minor fixes in view of DataSetSerialiser

* getDataLabel(int) returning 'null' if undefined, (the fall-back being
now implemented/moved to the DataToolTip plugin)
* Replaced 'Collections.<String>emptyList();' with ArrayList
implementation (as lightweight and can be equally overwritten if
necessary)

* Added prototype MultiArray Implementation

* Added PMD suppressions for unavoidable NPath complexity necessary for
handling of all primitive types.

* Removed unused variable

* Changed implicit cast to explicit (float) cast

* Changed implicit 'float' cast to an explicite one

* Fixed incomparable DecimalFormatter to MyDecimalFormatter '!equal! check

* Fixed a (too) early return in 'default' case bug

* Added missing hashCode() function

* Added computeIfAbsent and removed dead/commented code

* Fixed simple a==b for NaN statement

* Fixed minor HashMap cache invalidator

* Added missing transfer of info, warning and error meta info messages

* Fixed wrong NaN comparison

* Fixed array length issue

* Fixed double[] to a string formatting issue

* Changed implicit conversion to int to double (no conversion)

* Fixed explicit call to this (better than super)

* Removed the superfluous check

* Removed dead code

* Added explicit super and simplified boolean statement

* Replaced superfluous boxing Integer type with int

* Fixed code formatting and variable naming errors

* Simplified casting

* Fixed logical statement

* Fixed string format

* suppressed PMD complexity (seems to be unavoidable)

* suppressed PMD naming rule - class name is according to 'nomen est omen'

* added code-of-conduct and contributing descriptions

+++ chartfx-math/src/main/java/de/gsi/math/TRandom1.java
@@ -157,7 +157,7 @@ public class TRandom1 extends TRandom {
 
         uni = fFloatSeedTable[fJlag] - fFloatSeedTable[fIlag] - fCarry;
         if (uni < 0.) {
-            uni += 1.0;
+            uni += 1.0f;
             fCarry = (float) fMantissaBit24;
         } else {
             fCarry = 0.f;
@@ -172,7 +172,7 @@ public class TRandom1 extends TRandom {
             fJlag = 23;
 
         if (uni < fMantissaBit12) {
-            uni += fMantissaBit24 * fFloatSeedTable[fJlag];
+            uni += (float)fMantissaBit24 * fFloatSeedTable[fJlag];
             if (uni == 0)
                 uni = (float) (fMantissaBit24 * fMantissaBit24);
         }
@@ -187,7 +187,7 @@ public class TRandom1 extends TRandom {
             for (i = 0; i != fNskip; i++) {
                 uni = fFloatSeedTable[fJlag] - fFloatSeedTable[fIlag] - fCarry;
                 if (uni < 0.) {
-                    uni += 1.0;
+                    uni += 1.0f;
                     fCarry = (float) fMantissaBit24;
                 } else {
                     fCarry = 0.f;
@@ -228,7 +228,7 @@ public class TRandom1 extends TRandom {
         for (index = 0; index < size; ++index) {
             uni = fFloatSeedTable[fJlag] - fFloatSeedTable[fIlag] - fCarry;
             if (uni < 0.) {
-                uni += 1.0;
+                uni += 1.0f;
                 fCarry = (float) fMantissaBit24;
             } else {
                 fCarry = 0.f;
@@ -259,7 +259,7 @@ public class TRandom1 extends TRandom {
                 for (i = 0; i != fNskip; i++) {
                     uni = fFloatSeedTable[fJlag] - fFloatSeedTable[fIlag] - fCarry;
                     if (uni < 0.) {
-                        uni += 1.0;
+                        uni += 1.0f;
                         fCarry = (float) fMantissaBit24;
                     } else {
                         fCarry = 0.f;
